PP-11121 refactor service tests names
- Rename service tests that are for initiating pact tests. This is so it's easy to find which tests are used for pact testing and which for functional testing.
