+++ src/test/java/uk/gov/pay/api/service/AuthorisationServicePactTest.java
@@ -27,7 +27,7 @@ import static org.junit.Assert.assertThrows;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public class AuthorisationServiceTest {
+public class AuthorisationServicePactTest {
     private AuthorisationService authorisationService;
     private AuthorisationRequest request;
     private AuthorisationRequestExceptionMapper mapper;
