+++ src/test/java/uk/gov/pay/api/service/GetPaymentServicePactTest.java
@@ -34,7 +34,7 @@ import static org.hamcrest.core.Is.is;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public class GetPaymentServiceTest {
+public class GetPaymentServicePactTest {
 
     private static final String ACCOUNT_ID = "123456";
     private static final String CHARGE_ID = "ch_123abc456def";
